[Hackability Refactor] Move model.py underneath torchchat/ (#1077)
* Move model.py

* Clear out init to avoid package circular import
