+++ dist_run.py
@@ -11,7 +11,7 @@ import torch
 import torch.distributed as dist
 from torch.distributed.pipelining import PipelineStage, ScheduleGPipe
 
-from build.model import ModelArgs
+from torchchat.model import ModelArgs
 from build.model_dist import TransformerStage
 
 # Model config
