+++ eval.py
@@ -16,7 +16,7 @@ from build.builder import (
     TokenizerArgs,
 )
 
-from build.model import Model
+from torchchat.model import Model
 from build.utils import set_precision
 from cli import add_arguments_for_verb, arg_init
 from utils.measure_time import measure_time
