--- torchchat/__init__.py
@@ -1,3 +1,0 @@
-from torchchat import usages 
-
-__all__ = [usages]
